// src/types.ts
export interface Venue {
	id: number;
	app_site_id: number | null;
	name: string;
	physical_address: string | null;
	suburb: string | null;
	state: string | null;
	country: string | null;
	country_id: number | null;
	country_code: string | null;
	postal_address: string | null;
	telephone: string | null;
	website: string | null;
	email: string | null;
	fb_page_url: string | null;
	g_place_id: string;
	latitude: number | null;
	longitude: number | null;
	elevation: number | null;
	g_map_url: string | null;
	member_only: string | null;
	guest_welcome: string | null;
	pay_play: string | null;
	no_of_glass_courts: number | null;
	no_of_non_glass_courts: number | null;
	no_of_courts: number;
	no_of_doubles_courts: number | null;
	no_of_movable_walls: number | null;
	no_of_outdoor_courts: number | null;
	no_of_singles_courts: number | null;
	no_of_hardball_doubles_courts: number | null;
	membership_fees: string | null;
	court_rental_fees: string | null;
	peak_times: string | null;
	club_shop: string | null;
	sports_shop: string | null;
	hot_shower: string | null;
	gym: string | null;
	multi_sport_venue: string | null;
	cafe: string | null;
	serves_alcohol: string | null;
	water_fountain: string | null;
	drink_vending_machine: string | null;
	snack_vending_machine: string | null;
	free_parking: string | null;
	online_booking: string | null;
	twentyfour_hour_access: string | null;
	permanently_staffed: string | null;
	re_stringing_service: string | null;
	defibrillator: string | null;
	first_aid_kit: string | null;
	types: string | null;
	category: string | null;
	category_id: number | null;
	status: "0" | "1" | "2" | "3" | "4";
	reason_for_deletion: string | null;
	delete_reason_id: number | null;
	deletion_request_by_user_id: number | null;
	venue_deletion_reject_reason: string | null;
	more_details: string | null;
	business_status: string | null;
	reject_reason: string | null;
	reject_reason_id: number | null;
	additional_notes: string | null;
	membership_model: string | null;
	membership_numbers: string | null;
	non_member_numbers: string | null;
	collect_member_contact_information: string | null;
	most_profitable_area: string | null;
	biggest_frustration: string | null;
	support_from_sa_required: string | null;
	comments_notes: string | null;
	center_type: string | null;
	managers_name: string | null;
	insta_url: string | null;
	twitter_url: string | null;
	created_by: number | null;
	last_sync_at: Date | null;
	google_sync_status: number;
	date_approved: Date | null;
	date_flagged_for_deletion: Date | null;
	date_deleted: Date | null;
	nearest_venue_id: number;
	nearest_venue_km: number;
	nearest_venue_datetime_calc: Date | null;
	name_unique: string | null;
	name_normalised: string | null;
	created_at: Date;
	updated_at: Date;
}